+++ apps/analog-app/src/app/app.config.ts
@@ -5,14 +5,18 @@ import {
 } from '@angular/common/http';
 import { ApplicationConfig } from '@angular/core';
 import { provideClientHydration } from '@angular/platform-browser';
-import { provideFileRouter } from '@analogjs/router';
+import { provideFileRouter, requestContextInterceptor } from '@analogjs/router';
 import { withNavigationErrorHandler } from '@angular/router';
+
 import { cookieInterceptor } from './interceptors/cookies.interceptor';
 
 export const appConfig: ApplicationConfig = {
   providers: [
     provideFileRouter(withNavigationErrorHandler(console.error)),
-    provideHttpClient(withFetch(), withInterceptors([cookieInterceptor])),
+    provideHttpClient(
+      withFetch(),
+      withInterceptors([cookieInterceptor, requestContextInterceptor])
+    ),
     provideClientHydration(),
   ],
 };