+++ apps/analog-app/src/app/cart.service.ts
@@ -7,7 +7,6 @@ import { Product } from './products';
 })
 export class CartService {
   items: Product[] = [];
-  private readonly apiURL = import.meta.env.VITE_ANALOG_PUBLIC_BASE_URL;
 
   constructor(private http: HttpClient) {}
 
@@ -26,7 +25,7 @@ export class CartService {
 
   getShippingPrices() {
     return this.http.get<{ type: string; price: number }[]>(
-      `${this.apiURL}/assets/shipping.json`
+      `/assets/shipping.json`
     );
   }
 }