+++ apps/analog-app/src/main.server.ts
@@ -1,7 +1,9 @@
 import 'zone.js/node';
-import { InjectionToken, enableProdMode } from '@angular/core';
+import { enableProdMode } from '@angular/core';
 import { bootstrapApplication } from '@angular/platform-browser';
 import { renderApplication } from '@angular/platform-server';
+import { provideServerContext } from '@analogjs/router/server';
+import { ServerContext } from '@analogjs/router/tokens';
 
 import { config } from './app/app.config.server';
 import { AppComponent } from './app/app.component';
@@ -10,9 +12,6 @@ if (import.meta.env.PROD) {
   enableProdMode();
 }
 
-const REQUEST = new InjectionToken<Request>('REQUEST');
-const RESPONSE = new InjectionToken<Response>('RESPONSE');
-
 export function bootstrap() {
   return bootstrapApplication(AppComponent, config);
 }
@@ -20,15 +19,12 @@ export function bootstrap() {
 export default async function render(
   url: string,
   document: string,
-  { req, res }: { req: Request; res: Response }
+  serverContext: ServerContext
 ) {
   const html = await renderApplication(bootstrap, {
     document,
     url,
-    platformProviders: [
-      { provide: REQUEST, useValue: req },
-      { provide: RESPONSE, useValue: res },
-    ],
+    platformProviders: [provideServerContext(serverContext)],
   });
 
   return html;