+++ packages/platform/src/lib/router-plugin.ts
@@ -1,5 +1,8 @@
-import { VERSION } from '@angular/compiler-cli';
-import { normalizePath, Plugin } from 'vite';
+import { normalizePath, Plugin, UserConfig } from 'vite';
+import fg from 'fast-glob';
+import { resolve } from 'node:path';
+
+import { Options } from './options.js';
 
 /**
  * This plugin invalidates the files for routes when new files
@@ -9,52 +12,19 @@ import { normalizePath, Plugin } from 'vite';
  *
  * @returns
  */
-export function routerPlugin(): Plugin[] {
+export function routerPlugin(options?: Options): Plugin[] {
+  const workspaceRoot = options?.workspaceRoot ?? process.cwd();
+  let config: UserConfig;
+  let root: string;
+
   return [
-    {
-      name: 'analogjs-router-plugin',
-      config() {
-        return {
-          ssr: {
-            noExternal: [
-              '@analogjs/**',
-              '@analogjs/trpc/**',
-              '@angular/**',
-              '@angular/cdk/**',
-              '@angular/fire/**',
-              '@ngrx/**',
-              '@rx-angular/**',
-              '@ng-bootstrap/**',
-              '@ngneat/**',
-              'apollo-angular/**',
-              'primeng/**',
-              'rxfire/**',
-              '@tanstack/**',
-              'ngxtension/**',
-              'firebase/**',
-              'firebase-admin/**',
-            ],
-          },
-          optimizeDeps: {
-            include: [
-              '@angular/common',
-              '@angular/common/http',
-              ...(Number(VERSION.major) > 15
-                ? ['@angular/core/rxjs-interop']
-                : []),
-            ],
-            exclude: ['@angular/platform-server', '@analogjs/router'],
-          },
-        };
-      },
-    },
     {
       name: 'analogjs-router-invalidate-routes',
       configureServer(server) {
         function invalidateRoutes(path: string) {
           if (
             path.includes(normalizePath(`/app/routes/`)) ||
-            path.includes(normalizePath(`/app/pages/`))
+            path.includes(normalizePath(`/pages/`))
           ) {
             server.moduleGraph.fileToModulesMap.forEach((mods) => {
               mods.forEach((mod) => {
@@ -79,6 +49,107 @@ export function routerPlugin(): Plugin[] {
 
         server.watcher.on('add', invalidateRoutes);
         server.watcher.on('unlink', invalidateRoutes);
+      },
+    },
+    {
+      name: 'analog-glob-routes',
+      config(_config) {
+        config = _config;
+        root = resolve(workspaceRoot, config.root || '.') || '.';
+      },
+      transform(code) {
+        if (
+          code.includes('ANALOG_ROUTE_FILES') ||
+          code.includes('ANALOG_CONTENT_ROUTE_FILES')
+        ) {
+          const routeFiles: string[] = fg.sync(
+            [
+              `${root}/app/routes/**/*.ts`,
+              `${root}/src/app/routes/**/*.ts`,
+              `${root}/src/app/pages/**/*.page.ts`,
+              `${root}/src/app/pages/**/*.page.analog`,
+              ...(options?.additionalPagesDirs || [])?.map(
+                (glob) => `${workspaceRoot}${glob}/**/*.page.ts`
+              ),
+            ],
+            { dot: true }
+          );
+
+          const contentRouteFiles: string[] = fg.sync(
+            [
+              `${root}/src/app/routes/**/*.md`,
+              `${root}/src/app/pages/**/*.md`,
+              `${root}/src/content/**/*.md`,
+              ...(options?.additionalContentDirs || [])?.map(
+                (glob) => `${workspaceRoot}${glob}/**/*.md`
+              ),
+            ],
+            { dot: true }
+          );
+
+          let result = code.replace(
+            'let ANALOG_ROUTE_FILES = {};',
+            `
+            let ANALOG_ROUTE_FILES = {${routeFiles.map(
+              (module) =>
+                `"${module.replace(root, '')}": () => import('${module}')`
+            )}};
+          `
+          );
+
+          result = result.replace(
+            'let ANALOG_CONTENT_ROUTE_FILES = {};',
+            `
+          let ANALOG_CONTENT_ROUTE_FILES = {${contentRouteFiles.map(
+            (module) =>
+              `"${module.replace(
+                root,
+                ''
+              )}": () => import('${module}?analog-content-file=true').then(m => m.default)`
+          )}};
+          `
+          );
+
+          return {
+            code: result,
+            map: { mappings: '' },
+          };
+        }
+
+        return;
+      },
+    },
+    {
+      name: 'analog-glob-endpoints',
+      transform(code) {
+        if (code.includes('ANALOG_PAGE_ENDPOINTS')) {
+          const endpointFiles: string[] = fg.sync(
+            [
+              `${root}/src/app/pages/**/*.server.ts`,
+              ...(options?.additionalPagesDirs || [])?.map(
+                (glob) => `${workspaceRoot}${glob}/**/*.server.ts`
+              ),
+            ],
+            { dot: true }
+          );
+
+          const result = code.replace(
+            'let ANALOG_PAGE_ENDPOINTS = {};',
+            `
+            let ANALOG_PAGE_ENDPOINTS = {${endpointFiles.map(
+              (module) =>
+                `"${module.replace(root, '')}": () => import('${module}')`
+            )}};
+          `
+          );
+
+          return {
+            code: result,
+            map: { mappings: '' },
+          };
+        }
+
+        return;
       },
     },
   ];