+++ packages/router/src/index.ts
@@ -12,3 +12,4 @@ export { MetaTag } from './lib/meta-tags';
 export { PageServerLoad, LoadResult } from './lib/route-types';
 export { injectLoad } from './lib/inject-load';
 export { getLoadResolver } from './lib/get-load-resolver';
+export { requestContextInterceptor } from './lib/request-context';